Move getrusage() calls from dowait() to $&time
