+++ es.h
@@ -226,8 +226,8 @@ extern int efork(Boolean parent, Boolean background);
 extern pid_t spgrp(pid_t pgid);
 extern int tctakepgrp(void);
 extern void initpgrp(void);
-extern int ewait(int pid, Boolean interruptible, void *rusage);
-#define	ewaitfor(pid)	ewait(pid, FALSE, NULL)
+extern int ewait(int pid, Boolean interruptible);
+#define	ewaitfor(pid)	ewait(pid, FALSE)
 
 #if JOB_PROTECT
 extern void tcreturnpgrp(void);
